Organized .mjs files into auth and toolbox folders
