+++ server/routes/toolbox/inventory.mjs
